+++ src/neuroagent/app/app_utils.py
@@ -1,8 +1,17 @@
 """App utilities functions."""
 
+import logging
+from typing import Any
+
 from fastapi import HTTPException
 from httpx import AsyncClient
+from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
 from starlette.status import HTTP_401_UNAUTHORIZED
+
+from neuroagent.app.config import Settings
+
+logger = logging.getLogger(__name__)
 
 
 async def validate_project(
@@ -22,3 +31,31 @@ async def validate_project(
             status_code=HTTP_401_UNAUTHORIZED,
             detail="User does not belong to the project.",
         )
+
+
+def setup_engine(
+    settings: Settings, connection_string: str | None = None
+) -> AsyncEngine | None:
+    """Get the SQL engine."""
+    if connection_string:
+        engine_kwargs: dict[str, Any] = {"url": connection_string}
+        if "sqlite" in settings.db.prefix:  # type: ignore
+            # https://fastapi.tiangolo.com/tutorial/sql-databases/#create-the-sqlalchemy-engine
+            engine_kwargs["connect_args"] = {"check_same_thread": False}
+        engine = create_async_engine(**engine_kwargs)
+    else:
+        logger.warning("The SQL db_prefix needs to be set to use the SQL DB.")
+        return None
+    try:
+        engine.connect()
+        logger.info(
+            "Successfully connected to the SQL database"
+            f" {connection_string if not settings.db.password else connection_string.replace(settings.db.password.get_secret_value(), '*****')}."
+        )
+        return engine
+    except SQLAlchemyError:
+        logger.warning(
+            "Failed connection to SQL database"
+            f" {connection_string if not settings.db.password else connection_string.replace(settings.db.password.get_secret_value(), '*****')}."
+        )
+        return None